+++ src/main/java/snw/rfm/commands/admin/StartCommand.java
@@ -18,7 +18,7 @@ import org.bukkit.command.CommandSender;
 import org.jetbrains.annotations.NotNull;
 import snw.rfm.RunForMoney;
 import snw.rfm.events.GameStartEvent;
-import snw.rfm.game.GameConfiguration;
+import snw.rfm.config.GameConfiguration;
 import snw.rfm.game.GameProcess;
 import snw.rfm.game.TeamHolder;
 import snw.rfm.tasks.HunterReleaseTimer;
@@ -36,13 +36,12 @@ public final class StartCommand implements CommandExecutor {
             } else if (holder.isNoRunnerFound()) {
                 sender.sendMessage(ChatColor.RED + "" + ChatColor.BOLD + "没有人在逃走队员队伍里，因此无法启动游戏。");
             } else {
-                GameConfiguration conf = GameConfiguration.getInstance();
                 int lh = holder.getHunters().toArray().length;
                 int lr = holder.getRunners().toArray().length;
                 // 2022/1/30 修复游戏人数检查不严谨的错误
-                if (conf.getLeastHunter() < lh) {
+                if (GameConfiguration.getLeastHunter() < lh) {
                     sender.sendMessage(ChatColor.RED + "猎人数量小于管理员设置的下限值。最少需要 " + lh + " 具。");
-                } else if (conf.getLeastRunner() < lr) {
+                } else if (GameConfiguration.getLeastRunner() < lr) {
                     sender.sendMessage(ChatColor.RED + "逃走队员数量小于管理员设置的下限值。最少需要 " + lr + " 人。");
                 } else {
                     Bukkit.getPluginManager().callEvent(new GameStartEvent());