+++ src/main/java/snw/rfm/commands/hunter/DeactivateHunterCommand.java
@@ -15,18 +15,19 @@ import org.bukkit.ChatColor;
 import org.bukkit.command.Command;
 import org.bukkit.command.CommandExecutor;
 import org.bukkit.command.CommandSender;
+import org.bukkit.command.TabCompleter;
 import org.bukkit.entity.Player;
 import org.jetbrains.annotations.NotNull;
+import org.jetbrains.annotations.Nullable;
 import snw.rfm.RunForMoney;
+import snw.rfm.Util;
 import snw.rfm.game.GameProcess;
 import snw.rfm.game.TeamHolder;
 
-import java.util.ArrayList;
-import java.util.Arrays;
-import java.util.HashSet;
-import java.util.Iterator;
+import java.util.*;
+import java.util.stream.Collectors;
 
-public final class DeactivateHunterCommand implements CommandExecutor {
+public final class DeactivateHunterCommand implements CommandExecutor, TabCompleter {
     @Override
     public boolean onCommand(@NotNull CommandSender sender, @NotNull Command command, @NotNull String label, @NotNull String[] args) {
         if (args.length == 0) {
@@ -72,12 +73,20 @@ public final class DeactivateHunterCommand implements CommandExecutor {
                                 break;
                             }
                         }
-                        sender.sendMessage(ChatColor.RED + "其中，有 " + failed.toArray().length + " 个玩家因为不存在而禁用失败。");
+                        sender.sendMessage(ChatColor.RED + "其中，有 " + failed.toArray().length + " 个玩家因为不存在或不是猎人而禁用失败。");
                         sender.sendMessage(ChatColor.RED + "禁用失败的有: " + stringBuilder);
                     }
                 }
             }
         }
         return true;
+    }
+
+
+
+    @Nullable
+    @Override
+    public List<String> onTabComplete(@NotNull CommandSender sender, @NotNull Command command, @NotNull String alias, @NotNull String[] args) {
+        return (args.length > 0) ? Util.getAllPlayersName().stream().filter(IT -> TeamHolder.getInstance().isHunter(Bukkit.getPlayerExact(IT))).filter(IT -> !Arrays.asList(args).contains(IT)).collect(Collectors.toList()) : null;
     }
 }