+++ src/main/java/snw/rfm/processor/HunterPauseCardItemProcessor.java
@@ -1,4 +1,4 @@
-/**
+/*
  * This file is part of RunForMoney.
  *
  * RunForMoney is free software: you can redistribute it and/or modify it under the terms of the GNU General Public License as published by the Free Software Foundation, either version 3 of the License, or (at your option) any later version.
@@ -8,7 +8,7 @@
  * You should have received a copy of the GNU General Public License along with RunForMoney. If not, see <https://www.gnu.org/licenses/>.
  */
 
-package snw.rfm.item.processor;
+package snw.rfm.processor;
 
 import org.bukkit.Bukkit;
 import org.bukkit.ChatColor;
@@ -23,8 +23,8 @@ import org.bukkit.potion.PotionEffect;
 import org.bukkit.potion.PotionEffectType;
 import snw.rfm.RunForMoney;
 import snw.rfm.game.TeamHolder;
-import snw.rfm.item.ItemConfiguration;
-import snw.rfm.item.RFMItems;
+import snw.rfm.config.ItemConfiguration;
+import snw.rfm.RFMItems;
 
 public final class HunterPauseCardItemProcessor implements Listener {
     @EventHandler(priority = EventPriority.LOWEST) // 不用 LOWEST 会导致获取不到事件，因为 Spigot 内部会自动取消右键空气事件，一般的优先级 (中等) 不行
@@ -41,7 +41,7 @@ public final class HunterPauseCardItemProcessor implements Listener {
             if (item.isSimilar(RFMItems.HUNTER_PAUSE_CARD)) {
                 TeamHolder th = TeamHolder.getInstance();
                 if (!th.isHunter(player)) { // 排除使用者是猎人从而导致猎人坑队友的情况，哈哈哈哈哈哈哈笑死我了
-                    int hpctime = ItemConfiguration.getInstance().getItemTime("hpc") * 20; // 2022/2/6 避免重复计算
+                    int hpctime = ItemConfiguration.getItemTime("hpc") * 20; // 2022/2/6 避免重复计算
                     Bukkit.broadcastMessage(ChatColor.GREEN + player.getName() + " 使用了猎人暂停卡。猎人暂停了 " + hpctime + " 秒。");
                     item.setAmount(item.getAmount() - 1); // 2022/2/3 v1.1.3 修复了猎人使用此道具无效还使道具消失的错误。
                     for (Player p : th.getHunters()) {